refactor: consolidate transport abstractions into transport/ package
- Create transport/tcp_rtu.py: full RawRtuOverTcpTransport (ported from
  modbus_transport_raw.py; imports updated to use transport/crc and transport/raw)
- Rewrite transport/rtu.py: full RtuModbusTransport extending _ClientBackedTransport
  from transport/tcp; replaces the old incomplete stub
- Update transport/__init__.py: canonical re-export of all transport public names
- Shrink modbus_transport_base.py to 30-line shim (imports from transport.base,
  transport.raw, transport.crc; keeps classify_transport_error tuple wrapper)
- Shrink modbus_transport_client.py to 14-line shim (imports from transport.tcp/.rtu)
- Shrink modbus_transport_raw.py to 5-line shim (imports from transport.tcp_rtu)
- Update modbus_transport.py: now a flat re-export of transport/ symbols; no logic
- Update production imports (coordinator/, scanner/) to use transport.* directly
- Update test imports from modbus_transport.* to transport.* canonical paths
- Add tests/test_transport_package.py proving canonical hierarchy and shim identity

Remaining compat files: modbus_transport.py, modbus_transport_base.py,
modbus_transport_client.py, modbus_transport_raw.py kept as shims because
test_error_contract.py imports the module by name and external HA dynamic
import paths may still reference these top-level names.
