+++ custom_components/thessla_green_modbus/_coordinator_transport_select.py
@@ -8,7 +8,7 @@ from collections.abc import Awaitable, Callable
 
 from .const import CONNECTION_MODE_TCP, CONNECTION_MODE_TCP_RTU, DEFAULT_PORT
 from .modbus_exceptions import ConnectionException, ModbusException, ModbusIOException
-from .modbus_transport import BaseModbusTransport
+from .transport.base import BaseModbusTransport
 
 
 async def select_auto_transport(
