+++ custom_components/thessla_green_modbus/coordinator/io.py
@@ -49,7 +49,7 @@ from .retry import (
 )
 
 if TYPE_CHECKING:
-    from ..modbus_transport import BaseModbusTransport
+    from ..transport.base import BaseModbusTransport
 
 
 class _ModbusIOMixin:
