+++ tests/test_modbus_transport_base.py
@@ -13,7 +13,7 @@ from custom_components.thessla_green_modbus.modbus_exceptions import (
     ModbusException,
     ModbusIOException,
 )
-from custom_components.thessla_green_modbus.modbus_transport import TcpModbusTransport
+from custom_components.thessla_green_modbus.transport.tcp import TcpModbusTransport
 
 
 def _make_tcp(connection_type=CONNECTION_TYPE_TCP, **kwargs):
