+++ tests/test_modbus_transport_lifecycle.py
@@ -7,7 +7,7 @@ from unittest.mock import AsyncMock, patch
 
 import pytest
 from custom_components.thessla_green_modbus.const import CONNECTION_TYPE_TCP
-from custom_components.thessla_green_modbus.modbus_transport import TcpModbusTransport
+from custom_components.thessla_green_modbus.transport.tcp import TcpModbusTransport
 
 
 def _make_tcp(connection_type=CONNECTION_TYPE_TCP, **kwargs):
