+++ tests/test_modbus_transport_tcp.py
@@ -11,7 +11,7 @@ from custom_components.thessla_green_modbus.const import (
     CONNECTION_TYPE_TCP_RTU,
 )
 from custom_components.thessla_green_modbus.modbus_exceptions import ConnectionException
-from custom_components.thessla_green_modbus.modbus_transport import TcpModbusTransport
+from custom_components.thessla_green_modbus.transport.tcp import TcpModbusTransport
 
 
 def _make_tcp(connection_type=CONNECTION_TYPE_TCP, **kwargs):
